move getUsd to own file
