+++ src/lib/swapper/OneInchSwapper/getTradeQuote/getTradeQuote.test.ts
@@ -1,11 +1,17 @@
 import axios from 'axios'
 
 import { setupQuote } from '../../../../../packages/swapper/src/swappers/utils/test-data/setupSwapQuote'
+import { MAX_ONEINCH_TRADE } from '../utils/constants'
+import { getRate } from '../utils/helpers'
 import type { OneInchSwapperDeps } from '../utils/types'
 import { getTradeQuote } from './getTradeQuote'
 
 jest.mock('axios')
 const mockAxios = axios as jest.Mocked<typeof axios>
+
+jest.mock('../getUsdRate/getUsdRate', () => ({
+  getUsdRate: () => '0.02000',
+}))
 
 describe('getTradeQuote', () => {
   const deps: OneInchSwapperDeps = {
@@ -63,5 +69,7 @@ describe('getTradeQuote', () => {
     const quote = await getTradeQuote(deps, quoteInput)
     expect(quote.rate).toBe('0.000016426735042245')
     expect(quote.allowanceContract).toBe('0x1111111254eeb25477b68fb85ed929f73a960583')
+    expect(quote.maximumCryptoHuman).toBe(MAX_ONEINCH_TRADE)
+    expect(quote.minimumCryptoHuman).toBe('50')
   })
 })
